Use upstream signature formatting code for pylsp (#5214)
## 📝 Summary

This is a follow-up to #4864 since the signature-formatting capability
was incorporated directly into `python-lsp-server`

## 🔍 Description of Changes

Removed custom plugin and typechecking dependency on
`python-lsp-server`, added configuration to prefer `ruff` if enabled
over `black` (default in `python-lsp-server`).

This should remain in draft pending a new release of
`python-lsp-server`.

## 📋 Checklist

- [x] I have read the [contributor
guidelines](https://github.com/marimo-team/marimo/blob/main/CONTRIBUTING.md).
- [ ] For large changes, or changes that affect the public API: this
change was discussed or approved through an issue, on
[Discord](https://marimo.io/discord?ref=pr), or the community
[discussions](https://github.com/marimo-team/marimo/discussions) (Please
provide a link if applicable).
- [ ] I have added tests for the changes made.
- [ ] I have run the code and verified that it works as expected.

## 📜 Reviewers

mscolnick

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>